Fix linter errors

+++ src/index.js
@@ -1,13 +1,19 @@
+/* eslint-disable import/extensions */
 import './style.css';
-import {newTask} from './modules/to-do-UI';
+import { newTask } from './modules/to-do-UI';
 import { deleteItem } from './modules/removeItem';
-import dynamicLI from './modules/dynamicUI'
+import dynamicLI from './modules/dynamicUI';
 
+const inputField = document.querySelector('.input');
+inputField.addEventListener('keydown', (event) => {
+  if (event.key === 'Enter') {
+    dynamicLI();
+    deleteItem();
+  } else {
+    return dynamicLI();
+  }
+});
 
-let tasks = JSON.parse(localStorage.getItem('tasks')) || [];
-
+newTask();
+deleteItem();
 dynamicLI();
-newTask(tasks);
-deleteItem()
-
-
